Split the services by the system
