+++ services/webb-relayer/src/service/mod.rs
@@ -27,67 +27,24 @@ use std::sync::Arc;
 
 use axum::routing::get;
 use axum::Router;
-use ethereum_types::U256;
 use tower_http::cors::Any;
 use tower_http::cors::CorsLayer;
 use tower_http::trace::TraceLayer;
-use webb::evm::ethers::providers;
-use webb::substrate::subxt::config::{PolkadotConfig, SubstrateConfig};
-use webb::substrate::subxt::{tx::PairSigner, OnlineClient};
-use webb_bridge_registry_backends::dkg::DkgBridgeRegistryBackend;
-use webb_bridge_registry_backends::mocked::MockedBridgeRegistryBackend;
-use webb_event_watcher_traits::evm::{BridgeWatcher, EventWatcher};
-use webb_event_watcher_traits::substrate::SubstrateBridgeWatcher;
-use webb_event_watcher_traits::SubstrateEventWatcher;
-use webb_ew_dkg::DKGMetadataWatcher;
-use webb_ew_dkg::DKGProposalHandlerWatcher;
-use webb_ew_dkg::DKGPublicKeyChangedHandler;
-use webb_ew_dkg::ProposalSignedHandler;
-use webb_ew_evm::open_vanchor::{
-    OpenVAnchorDepositHandler, OpenVAnchorLeavesHandler,
-};
-use webb_ew_evm::signature_bridge_watcher::{
-    SignatureBridgeContractWatcher, SignatureBridgeContractWrapper,
-    SignatureBridgeGovernanceOwnershipTransferredHandler,
-};
-use webb_ew_evm::vanchor::vanchor_encrypted_outputs_handler::VAnchorEncryptedOutputHandler;
-use webb_ew_evm::{
-    vanchor::*, OpenVAnchorContractWatcher, OpenVAnchorContractWrapper,
-    VAnchorContractWatcher, VAnchorContractWrapper,
-};
-use webb_ew_substrate::MaintainerSetEventHandler;
-use webb_ew_substrate::SubstrateBridgeEventWatcher;
-use webb_ew_substrate::SubstrateVAnchorEventWatcher;
-use webb_ew_substrate::{
-    SubstrateVAnchorDepositHandler, SubstrateVAnchorEncryptedOutputHandler,
-    SubstrateVAnchorLeavesHandler,
-};
+use webb::substrate::subxt;
 use webb_proposal_signing_backends::*;
 use webb_relayer_config::anchor::LinkedAnchorConfig;
-use webb_relayer_config::evm::{
-    Contract, SignatureBridgeContractConfig, VAnchorContractConfig,
-};
+
 use webb_relayer_config::signing_backend::ProposalSigningBackendConfig;
-use webb_relayer_config::substrate::{
-    DKGPalletConfig, DKGProposalHandlerPalletConfig, Pallet,
-    SignatureBridgePalletConfig, SubstrateRuntime, VAnchorBn254PalletConfig,
-};
 use webb_relayer_context::RelayerContext;
 use webb_relayer_handlers::routes::info::handle_relayer_info;
-use webb_relayer_handlers::routes::metric;
 use webb_relayer_handlers::{handle_socket_info, websocket_handler};
 use webb_relayer_store::SledStore;
-use webb_relayer_tx_queue::{evm::TxQueue, substrate::SubstrateTxQueue};
 
 /// EVM Specific Services
 pub mod evm;
 /// Substrate Specific Services
 pub mod substrate;
 
-/// Type alias for the DKG DefaultConfig
-pub type DkgClient = OnlineClient<PolkadotConfig>;
-/// Type alias for the WebbProtocol DefaultConfig
-pub type WebbProtocolClient = OnlineClient<SubstrateConfig>;
 /// Type alias for [Sled](https://sled.rs)-based database store
 pub type Store = SledStore;
 
@@ -134,603 +91,8 @@ pub async fn ignite(
         "Relayer configuration: {}",
         serde_json::to_string_pretty(&ctx.config)?
     );
-
     evm::ignite(ctx, store.clone()).await?;
-    for (node_name, node_config) in &ctx.config.substrate {
-        if !node_config.enabled {
-            continue;
-        }
-        let chain_id = node_config.chain_id;
-        match node_config.runtime {
-            SubstrateRuntime::Dkg => {
-                let client =
-                    ctx.substrate_provider::<PolkadotConfig>(node_name).await?;
-
-                let chain_id = chain_id;
-                for pallet in &node_config.pallets {
-                    match pallet {
-                        Pallet::DKGProposalHandler(config) => {
-                            start_dkg_proposal_handler(
-                                ctx,
-                                config,
-                                client.clone(),
-                                chain_id,
-                                store.clone(),
-                            )?;
-                        }
-                        Pallet::Dkg(config) => {
-                            start_dkg_pallet_watcher(
-                                ctx,
-                                config,
-                                client.clone(),
-                                chain_id,
-                                store.clone(),
-                            )?;
-                        }
-                        Pallet::DKGProposals(_) => {
-                            // TODO(@shekohex): start the dkg proposals service
-                        }
-                        Pallet::SignatureBridge(_) => {
-                            unreachable!()
-                        }
-                        Pallet::VAnchorBn254(_) => {
-                            unreachable!()
-                        }
-                    }
-                }
-                // start the transaction queue for dkg-substrate extrinsics after starting other tasks.
-                start_dkg_substrate_tx_queue(
-                    ctx.clone(),
-                    chain_id,
-                    store.clone(),
-                )?;
-            }
-            SubstrateRuntime::WebbProtocol => {
-                let client = ctx
-                    .substrate_provider::<SubstrateConfig>(node_name)
-                    .await?;
-                let chain_id = chain_id;
-                for pallet in &node_config.pallets {
-                    match pallet {
-                        Pallet::VAnchorBn254(config) => {
-                            start_substrate_vanchor_event_watcher(
-                                ctx,
-                                config,
-                                client.clone(),
-                                chain_id,
-                                store.clone(),
-                            )?;
-                        }
-                        Pallet::SignatureBridge(config) => {
-                            start_substrate_signature_bridge_events_watcher(
-                                ctx.clone(),
-                                config,
-                                client.clone(),
-                                chain_id,
-                                store.clone(),
-                            )
-                            .await?;
-                        }
-                        Pallet::DKGProposals(_) => {
-                            unreachable!()
-                        }
-                        Pallet::DKGProposalHandler(_) => {
-                            unreachable!()
-                        }
-                        Pallet::Dkg(_) => {
-                            unreachable!()
-                        }
-                    }
-                }
-
-                // start the transaction queue for protocol-substrate  after starting other tasks.
-                start_protocol_substrate_tx_queue(
-                    ctx.clone(),
-                    chain_id,
-                    store.clone(),
-                )?;
-            }
-        };
-    }
-    Ok(())
-}
-
-/// Starts the event watcher for Substrate vanchor events.
-///
-/// Returns Ok(()) if successful, or an error if not.
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `config` - VAnchorBn254 configuration
-/// * `client` - WebbProtocol client
-/// * `chain_id` - An u32 representing the chain id of the chain
-/// * `store` -[Sled](https://sled.rs)-based database store
-pub fn start_substrate_vanchor_event_watcher(
-    ctx: &RelayerContext,
-    config: &VAnchorBn254PalletConfig,
-    client: WebbProtocolClient,
-    chain_id: u32,
-    store: Arc<Store>,
-) -> crate::Result<()> {
-    if !config.events_watcher.enabled {
-        tracing::warn!(
-            "Substrate VAnchor events watcher is disabled for ({}).",
-            chain_id,
-        );
-        return Ok(());
-    }
-    tracing::debug!(
-        "Substrate VAnchor events watcher for ({}) Started.",
-        chain_id,
-    );
-    // now, we start substrate service/tasks
-
-    let my_ctx = ctx.clone();
-    let my_config = config.clone();
-    let mut shutdown_signal = ctx.shutdown_signal();
-    let metrics = ctx.metrics.clone();
-    let task = async move {
-        let proposal_signing_backend = make_substrate_proposal_signing_backend(
-            &my_ctx,
-            store.clone(),
-            chain_id,
-            my_config.linked_anchors.clone(),
-            my_config.proposal_signing_backend,
-        )
-        .await?;
-        match proposal_signing_backend {
-            ProposalSigningBackendSelector::Dkg(backend) => {
-                // its safe to use unwrap on linked_anchors here
-                // since this option is always going to return Some(value).
-                // linked_anchors are validated in make_proposal_signing_backend() method
-                let deposit_handler = SubstrateVAnchorDepositHandler::new(
-                    backend,
-                    my_config.linked_anchors.unwrap(),
-                );
-                let leaves_handler = SubstrateVAnchorLeavesHandler::default();
-                let encrypted_output_handler =
-                    SubstrateVAnchorEncryptedOutputHandler::default();
-
-                let watcher = SubstrateVAnchorEventWatcher::default();
-                let substrate_vanchor_watcher_task = watcher.run(
-                    chain_id,
-                    client.clone().into(),
-                    store.clone(),
-                    my_config.events_watcher,
-                    vec![
-                        Box::new(deposit_handler),
-                        Box::new(leaves_handler),
-                        Box::new(encrypted_output_handler),
-                    ],
-                    metrics.clone(),
-                );
-
-                tokio::select! {
-                    _ = substrate_vanchor_watcher_task => {
-                        tracing::warn!(
-                            "Substrate VAnchor watcher (DKG Backend) task stopped for ({})",
-                            chain_id,
-                        );
-                    },
-
-                    _ = shutdown_signal.recv() => {
-                        tracing::trace!(
-                            "Stopping Substrate VAnchor watcher (DKG Backend) for ({})",
-                            chain_id,
-                        );
-                    },
-                }
-            }
-            ProposalSigningBackendSelector::Mocked(backend) => {
-                // its safe to use unwrap on linked_anchors here
-                // since this option is always going to return Some(value).
-                let deposit_handler = SubstrateVAnchorDepositHandler::new(
-                    backend,
-                    my_config.linked_anchors.unwrap(),
-                );
-                let leaves_handler = SubstrateVAnchorLeavesHandler::default();
-                let encrypted_output_handler =
-                    SubstrateVAnchorEncryptedOutputHandler::default();
-
-                let watcher = SubstrateVAnchorEventWatcher::default();
-                let substrate_vanchor_watcher_task = watcher.run(
-                    chain_id,
-                    client.clone().into(),
-                    store.clone(),
-                    my_config.events_watcher,
-                    vec![
-                        Box::new(deposit_handler),
-                        Box::new(leaves_handler),
-                        Box::new(encrypted_output_handler),
-                    ],
-                    metrics.clone(),
-                );
-                tokio::select! {
-                    _ = substrate_vanchor_watcher_task => {
-                        tracing::warn!(
-                            "Substrate VAnchor watcher (Mocked Backend) task stopped for ({})",
-                            chain_id,
-                        );
-                    },
-                    _ = shutdown_signal.recv() => {
-                        tracing::trace!(
-                            "Stopping Substrate VAnchor watcher (Mocked Backend) for ({})",
-                            chain_id,
-                        );
-                    },
-                }
-            }
-            ProposalSigningBackendSelector::None => {
-                let leaves_handler = SubstrateVAnchorLeavesHandler::default();
-                let encrypted_output_handler =
-                    SubstrateVAnchorEncryptedOutputHandler::default();
-
-                let watcher = SubstrateVAnchorEventWatcher::default();
-                let substrate_vanchor_watcher_task = watcher.run(
-                    chain_id,
-                    client.clone().into(),
-                    store.clone(),
-                    my_config.events_watcher,
-                    vec![
-                        Box::new(leaves_handler),
-                        Box::new(encrypted_output_handler),
-                    ],
-                    metrics.clone(),
-                );
-                tokio::select! {
-                    _ = substrate_vanchor_watcher_task => {
-                        tracing::warn!(
-                            "Substrate VAnchor watcher task stopped for ({})",
-                            chain_id,
-                        );
-                    },
-                    _ = shutdown_signal.recv() => {
-                        tracing::trace!(
-                            "Stopping Substrate VAnchor watcher task for ({})",
-                            chain_id,
-                        );
-                    },
-                }
-            }
-        };
-        crate::Result::Ok(())
-    };
-    // kick off the watcher.
-    tokio::task::spawn(task);
-    Ok(())
-}
-
-/// Starts the event watcher for DKG proposal handler events.
-///
-/// Returns Ok(()) if successful, or an error if not.
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `config` - DKG proposal handler configuration
-/// * `client` - DKG client
-/// * `chain_id` - An u32 representing the chain id of the chain
-/// * `store` -[Sled](https://sled.rs)-based database store
-pub fn start_dkg_proposal_handler(
-    ctx: &RelayerContext,
-    config: &DKGProposalHandlerPalletConfig,
-    client: DkgClient,
-    chain_id: u32,
-    store: Arc<Store>,
-) -> crate::Result<()> {
-    // check first if we should start the events watcher for this contract.
-    if !config.events_watcher.enabled {
-        tracing::warn!(
-            "DKG Proposal Handler events watcher is disabled for ({}).",
-            chain_id,
-        );
-        return Ok(());
-    }
-    tracing::debug!(
-        "DKG Proposal Handler events watcher for ({}) Started.",
-        chain_id,
-    );
-    let mut shutdown_signal = ctx.shutdown_signal();
-    let metrics = ctx.metrics.clone();
-    let my_config = config.clone();
-    let task = async move {
-        let proposal_handler_watcher = DKGProposalHandlerWatcher::default();
-        let proposal_signed_handler = ProposalSignedHandler::default();
-        let proposal_handler_watcher_task = proposal_handler_watcher.run(
-            chain_id,
-            client.into(),
-            store,
-            my_config.events_watcher,
-            vec![Box::new(proposal_signed_handler)],
-            metrics,
-        );
-        tokio::select! {
-            _ = proposal_handler_watcher_task => {
-                tracing::warn!(
-                    "DKG Proposal Handler events watcher stopped for ({})",
-                    chain_id,
-                );
-            },
-            _ = shutdown_signal.recv() => {
-                tracing::trace!(
-                    "Stopping DKG Proposal Handler events watcher for ({})",
-                    chain_id,
-                );
-            },
-        }
-    };
-    // kick off the watcher.
-    tokio::task::spawn(task);
-    Ok(())
-}
-
-/// Starts the event watcher for DKG pallet events watcher.
-///
-/// Returns Ok(()) if successful, or an error if not.
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `config` - DKG pallet configuration
-/// * `client` - DKG client
-/// * `chain_id` - An u32 representing the chain id of the chain
-/// * `store` -[Sled](https://sled.rs)-based database store
-pub fn start_dkg_pallet_watcher(
-    ctx: &RelayerContext,
-    config: &DKGPalletConfig,
-    client: DkgClient,
-    chain_id: u32,
-    store: Arc<Store>,
-) -> crate::Result<()> {
-    // check first if we should start the events watcher for this pallet.
-    if !config.events_watcher.enabled {
-        tracing::warn!(
-            "DKG Pallet events watcher is disabled for ({}).",
-            chain_id,
-        );
-        return Ok(());
-    }
-    tracing::debug!("DKG Pallet events watcher for ({}) Started.", chain_id,);
-    let mut shutdown_signal = ctx.shutdown_signal();
-    let webb_config = ctx.config.clone();
-    let metrics = ctx.metrics.clone();
-    let my_config = config.clone();
-    let task = async move {
-        let dkg_event_watcher = DKGMetadataWatcher::default();
-        let public_key_changed_handler =
-            DKGPublicKeyChangedHandler::new(webb_config);
-
-        let dkg_event_watcher_task = dkg_event_watcher.run(
-            chain_id,
-            client.into(),
-            store,
-            my_config.events_watcher,
-            vec![Box::new(public_key_changed_handler)],
-            metrics,
-        );
-        tokio::select! {
-            _ = dkg_event_watcher_task => {
-                tracing::warn!(
-                    "DKG Pallet events watcher stopped for ({})",
-                    chain_id,
-                );
-            },
-            _ = shutdown_signal.recv() => {
-                tracing::trace!(
-                    "Stopping DKG Pallet events watcher for ({})",
-                    chain_id,
-                );
-            },
-        }
-    };
-    // kick off the watcher.
-    tokio::task::spawn(task);
-    Ok(())
-}
-
-/// Starts the event watcher for Signature Bridge Pallet.
-pub async fn start_substrate_signature_bridge_events_watcher(
-    ctx: RelayerContext,
-    config: &SignatureBridgePalletConfig,
-    client: WebbProtocolClient,
-    chain_id: u32,
-    store: Arc<Store>,
-) -> crate::Result<()> {
-    if !config.events_watcher.enabled {
-        tracing::warn!(
-            "Substrate Signature Bridge events watcher is disabled for ({}).",
-            chain_id,
-        );
-        return Ok(());
-    }
-    let mut shutdown_signal = ctx.shutdown_signal();
-    let my_config = config.clone();
-    let task = async move {
-        tracing::debug!(
-            "Substrate Signature Bridge watcher for ({}) Started.",
-            chain_id
-        );
-        let substrate_bridge_watcher = SubstrateBridgeEventWatcher::default();
-        let bridge_event_handler = MaintainerSetEventHandler::default();
-        let events_watcher_task = SubstrateEventWatcher::run(
-            &substrate_bridge_watcher,
-            chain_id,
-            client.clone().into(),
-            store.clone(),
-            my_config.events_watcher,
-            vec![Box::new(bridge_event_handler)],
-            ctx.metrics.clone(),
-        );
-        let cmd_handler_task = SubstrateBridgeWatcher::run(
-            &substrate_bridge_watcher,
-            chain_id,
-            client.into(),
-            store.clone(),
-        );
-        tokio::select! {
-            _ = events_watcher_task => {
-                tracing::warn!(
-                    "Substrate signature bridge events watcher task stopped for ({})",
-                    chain_id
-                );
-            },
-            _ = cmd_handler_task => {
-                tracing::warn!(
-                    "Substrate signature bridge cmd handler task stopped for ({})",
-                    chain_id
-                );
-            },
-            _ = shutdown_signal.recv() => {
-                tracing::trace!(
-                    "Stopping Substrate Signature Bridge watcher for ({})",
-                    chain_id,
-                );
-            },
-        }
-    };
-    // kick off the watcher.
-    tokio::task::spawn(task);
-    Ok(())
-}
-
-/// Starts the transaction queue task
-///
-/// Returns Ok(()) if successful, or an error if not.
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `chain_name` - Name of the chain
-/// * `store` -[Sled](https://sled.rs)-based database store
-pub fn start_tx_queue(
-    ctx: RelayerContext,
-    chain_id: String,
-    store: Arc<Store>,
-) -> crate::Result<()> {
-    // Start tx_queue only when governance relaying feature is enabled for relayer.
-    if !ctx.config.features.governance_relay {
-        tracing::warn!("Tx Queue disabled for ({})", chain_id,);
-        return Ok(());
-    }
-
-    let mut shutdown_signal = ctx.shutdown_signal();
-    let tx_queue = TxQueue::new(ctx, chain_id.clone(), store);
-
-    tracing::debug!("Transaction Queue for ({}) Started.", chain_id);
-    let task = async move {
-        tokio::select! {
-            _ = tx_queue.run() => {
-                tracing::warn!(
-                    "Transaction Queue task stopped for ({})",
-                    chain_id,
-                );
-            },
-            _ = shutdown_signal.recv() => {
-                tracing::trace!(
-                    "Stopping Transaction Queue for ({})",
-                    chain_id,
-                );
-            },
-        }
-    };
-    // kick off the tx_queue.
-    tokio::task::spawn(task);
-    Ok(())
-}
-
-/// Starts the transaction queue task for protocol-substrate extrinsic
-///
-/// Returns Ok(()) if successful, or an error if not.
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `chain_name` - Name of the chain
-/// * `store` -[Sled](https://sled.rs)-based database store
-pub fn start_protocol_substrate_tx_queue(
-    ctx: RelayerContext,
-    chain_id: u32,
-    store: Arc<Store>,
-) -> crate::Result<()> {
-    // Start tx_queue only when governance relaying feature is enabled for relayer.
-    if !ctx.config.features.governance_relay {
-        tracing::warn!(
-            "Tx Queue disabled for Protocol-Substrate node ({})",
-            chain_id,
-        );
-        return Ok(());
-    }
-
-    let mut shutdown_signal = ctx.shutdown_signal();
-
-    let tx_queue = SubstrateTxQueue::new(ctx, chain_id, store);
-
-    tracing::debug!(
-        "Transaction Queue for Protocol-Substrate node({}) Started.",
-        chain_id
-    );
-    let task = async move {
-        tokio::select! {
-            _ = tx_queue.run::<SubstrateConfig>() => {
-                tracing::warn!(
-                    "Transaction Queue task stopped for Protocol-Substrate node({})",
-                    chain_id
-                );
-            },
-            _ = shutdown_signal.recv() => {
-                tracing::trace!(
-                    "Stopping Transaction Queue for Protocol-Substrate node({})",
-                    chain_id
-                );
-            },
-        }
-    };
-    // kick off the substrate tx_queue.
-    tokio::task::spawn(task);
-    Ok(())
-}
-
-/// Starts the transaction queue task for dkg-substrate extrinsics
-///
-/// Returns Ok(()) if successful, or an error if not.
-///
-/// # Arguments
-///
-/// * `ctx` - RelayContext reference that holds the configuration
-/// * `chain_name` - Name of the chain
-/// * `store` -[Sled](https://sled.rs)-based database store
-pub fn start_dkg_substrate_tx_queue(
-    ctx: RelayerContext,
-    chain_id: u32,
-    store: Arc<Store>,
-) -> crate::Result<()> {
-    let mut shutdown_signal = ctx.shutdown_signal();
-
-    let tx_queue = SubstrateTxQueue::new(ctx, chain_id, store);
-
-    tracing::debug!(
-        "Transaction Queue for Dkg-Substrate node({}) Started.",
-        chain_id
-    );
-    let task = async move {
-        tokio::select! {
-            _ = tx_queue.run::<PolkadotConfig>() => {
-                tracing::warn!(
-                    "Transaction Queue task stopped for Dkg-Substrate node({})",
-                    chain_id
-                );
-            },
-            _ = shutdown_signal.recv() => {
-                tracing::trace!(
-                    "Stopping Transaction Queue for Dkg-Substrate node({})",
-                    chain_id
-                );
-            },
-        }
-    };
-    // kick off the substrate tx_queue.
-    tokio::task::spawn(task);
+    substrate::ignite(ctx, store.clone()).await?;
     Ok(())
 }
 
@@ -748,7 +110,7 @@ pub enum ProposalSigningBackendSelector {
 pub async fn make_proposal_signing_backend(
     ctx: &RelayerContext,
     store: Arc<Store>,
-    chain_id: u32,
+    typed_chain_id: webb_proposals::TypedChainId,
     linked_anchors: Option<Vec<LinkedAnchorConfig>>,
     proposal_signing_backend: Option<ProposalSigningBackendConfig>,
 ) -> crate::Result<ProposalSigningBackendSelector> {
@@ -763,13 +125,13 @@ pub async fn make_proposal_signing_backend(
         Some(ProposalSigningBackendConfig::DkgNode(c)) => {
             // if it is the dkg backend, we will need to connect to that node first,
             // and then use the DkgProposalSigningBackend to sign the proposal.
-            let typed_chain_id = webb_proposals::TypedChainId::Evm(chain_id);
-            let dkg_client =
-                ctx.substrate_provider::<PolkadotConfig>(&c.node).await?;
+            let dkg_client = ctx
+                .substrate_provider::<subxt::PolkadotConfig>(&c.node)
+                .await?;
             let pair = ctx.substrate_wallet(&c.node).await?;
             let backend = DkgProposalSigningBackend::new(
                 dkg_client,
-                PairSigner::new(pair),
+                subxt::tx::PairSigner::new(pair),
                 typed_chain_id,
             );
             Ok(ProposalSigningBackendSelector::Dkg(backend))
@@ -807,83 +169,6 @@ pub async fn make_proposal_signing_backend(
                 };
                 signature_bridges.insert(resource_id);
             });
-            let backend = MockedProposalSigningBackend::builder()
-                .store(store.clone())
-                .private_key(mocked.private_key)
-                .signature_bridges(signature_bridges)
-                .build();
-            Ok(ProposalSigningBackendSelector::Mocked(backend))
-        }
-        None => {
-            tracing::warn!("Misconfigured Network: Proposal signing backend must be configured for governance relaying");
-            Ok(ProposalSigningBackendSelector::None)
-        }
-    }
-}
-
-/// utility to configure proposal signing backend for substrate chain
-pub async fn make_substrate_proposal_signing_backend(
-    ctx: &RelayerContext,
-    store: Arc<Store>,
-    chain_id: u32,
-    linked_anchors: Option<Vec<LinkedAnchorConfig>>,
-    proposal_signing_backend: Option<ProposalSigningBackendConfig>,
-) -> crate::Result<ProposalSigningBackendSelector> {
-    // Check if contract is configured with governance support for the relayer.
-    if !ctx.config.features.governance_relay {
-        tracing::warn!("Governance relaying is not enabled for relayer");
-        return Ok(ProposalSigningBackendSelector::None);
-    }
-    // check if linked anchors are provided.
-    let linked_anchors = match linked_anchors {
-        Some(anchors) => {
-            if anchors.is_empty() {
-                tracing::warn!("Misconfigured Network: Linked anchors cannot be empty for governance relaying");
-                return Ok(ProposalSigningBackendSelector::None);
-            } else {
-                anchors
-            }
-        }
-        None => {
-            tracing::warn!("Misconfigured Network: Linked anchors must be configured for governance relaying");
-            return Ok(ProposalSigningBackendSelector::None);
-        }
-    };
-    // we need to check/match on the proposal signing backend configured for this anchor.
-    match proposal_signing_backend {
-        Some(ProposalSigningBackendConfig::DkgNode(c)) => {
-            // if it is the dkg backend, we will need to connect to that node first,
-            // and then use the DkgProposalSigningBackend to sign the proposal.
-            let typed_chain_id =
-                webb_proposals::TypedChainId::Substrate(chain_id);
-            let dkg_client =
-                ctx.substrate_provider::<PolkadotConfig>(&c.node).await?;
-            let pair = ctx.substrate_wallet(&c.node).await?;
-            let backend = DkgProposalSigningBackend::new(
-                dkg_client,
-                PairSigner::new(pair),
-                typed_chain_id,
-            );
-            Ok(ProposalSigningBackendSelector::Dkg(backend))
-        }
-        Some(ProposalSigningBackendConfig::Mocked(mocked)) => {
-            // if it is the mocked backend, we will use the MockedProposalSigningBackend to sign the proposal.
-            // which is a bit simpler than the DkgProposalSigningBackend.
-            // get only the linked chains to that anchor.
-            let mut signature_bridges: HashSet<webb_proposals::ResourceId> =
-                HashSet::new();
-            linked_anchors.iter().for_each(|anchor| {
-                // using chain_id to ensure that we have only one signature bridge
-                let resource_id = match anchor {
-                    LinkedAnchorConfig::Raw(target) => {
-                        let bytes: [u8; 32] = target.resource_id.into();
-                        webb_proposals::ResourceId::from(bytes)
-                    }
-                    _ => unreachable!("unsupported"),
-                };
-                signature_bridges.insert(resource_id);
-            });
-
             let backend = MockedProposalSigningBackend::builder()
                 .store(store.clone())
                 .private_key(mocked.private_key)
